Finished site

+++ js/gameController.js
@@ -4,14 +4,17 @@
 *                         GAME                             *
 ***********************************************************/
 
-//Hide owner tools.
-//******Quick MVP solution. Major vulnerability.********
+
 $('#reportscore').hide();
+
+
 
 
 //Get game object
 var gameID = getQueryVariable("gameID");
 var game;
+
+
 
 //I'm making a lot of redundant calls to this database but Ill leave it for now. Lets fix it later.
 //This one gets the current game
@@ -31,14 +34,45 @@ currentGame.find({
         //populate other players
         var player = new Parse.Query(User);
 
-        if (game.get('player_id')) for (var i = 0; i < game.get('player_id').length; i++) {
+        var numPlayers = game.get('player_id').length;
+        
+        //expand box to fit all players
+        if (numPlayers > 4) {
+            var height = $('#game_box').height();
+            var linesToAdd = numPlayers - 4;
+            var newHeight = height + linesToAdd * 55;
+            $('#game_box').css("height", newHeight+"px");
+        }
+        var roster;
+        if (game.get('player_id')) for (var i = 0; i < numPlayers; i++) {
               player.equalTo("objectId", game.get('player_id')[i]);
               player.find({
                      success: function(players) {
                             var $playerElm = $('#playerstemplate').clone(true);
                             $playerElm.html(players[0].get('username')+" ("+Math.round(players[0].get('ranking'))+"pts)");
                             $playerElm.attr("id", players[0].id);
-                            $('#players').append($playerElm);
+                            $playerElm.attr("rank",  players[0].get('ranking'));
+                            var added = false;
+                            
+                            $('#players p').each(function(index) {
+                                if (added == false) {
+                                    console.log($(this).text());
+                                    if (index>0) {
+                                        var abc = players[0].get('ranking');
+                                        var def = $(this).attr('rank');
+                                    
+                                        if(players[0].get('ranking') > $(this).attr('rank')){
+                                            console.log("SORT ADDED");
+                                            $(this).before($playerElm);
+                                            added = true;
+                                        }
+                                    }
+                                }
+                            })
+                            
+                            if (added == false) {
+                                $('#players').append($playerElm)
+                            }
                      },
                      error: function(e) {
                             console.log(e);
@@ -49,6 +83,11 @@ currentGame.find({
               if (game.get('player_id')[i] == Parse.User.current().id) {
                 $('#joingame').html("LEAVE");
                 $('#joingame').attr("id", "leavegame");
+                
+                //REMOVE TOMORROW
+                $('#tournamentjoingame').html("LEAVE");
+                $('#tournamentjoingame').attr("id", "tournamentleavegame");
+                
               }
               
         }
@@ -70,7 +109,7 @@ Parse.User.current().fetch().then(function(user) {
     $('#rank').html(Math.round(user.get('ranking')));
 });
 
-$(document).on('click', '#joingame', function() {
+$(document).on('click', '#joingame, #tournamentjoingame', function() {
     
         console.log("JOINING GAME")
     
@@ -83,9 +122,11 @@ $(document).on('click', '#joingame', function() {
        
 });
 
-$(document).on('click', '#leavegame', function() {
+
+
+
+$(document).on('click', '#leavegame, #tournamentleavegame', function() {
     
-    console.log("LEAVING GAME")
     
     var currentPlayers = game.get('player_id');
     
@@ -125,6 +166,12 @@ $('#reportscore').click(function() {
     }
 });
 
+//REMOVE
+$('#tournamentreportscore').click(function() {
+    $('#scorebox').show();
+    $("html, body").animate({ scrollTop: $(document).height() }, "slow");
+});
+
 $('#scoreform').submit(function(event) {
     event.preventDefault();
     
@@ -137,12 +184,15 @@ $('#scoreform').submit(function(event) {
     //Text me if you need to change this. You need to have Parse command line tools installed
     Parse.Cloud.run('updateRankings', setRankingParams, {
         success: function(response) {
-            currentGame.get(gameID, {
-                success: function(tempGame) {
-                    tempGame.destroy({});
-                    window.location.href="./dashboard.html";
-                }
-            });
+            if (gameID != 'b5sfsVcln5') {
+                currentGame.get(gameID, {
+                    success: function(tempGame) {
+                        tempGame.destroy({});
+                        window.location.href="./dashboard.html";
+                    }
+                });
+            }
+            else location.reload();
         }
     });  
 })
